requirements.txt
python-telegram-bot==21.0
supabase==2.3.4
groq==0.4.2
python-dotenv==1.0.0
pytz==2024.1
